network: windows: Fix set_blocking_mode() nonblock type
ioctlsocket() expects a unsigned long rather than a unsigned int, so change
the type of the nonblock variable.

This is purely cosmetic since both types have the same size on Windows,
even on a 64-bit system.

+++ network.c
@@ -103,7 +103,7 @@ struct iio_device_pdata {
 
 static int set_blocking_mode(int s, bool blocking)
 {
-	unsigned int nonblock;
+	unsigned long nonblock;
 	int ret;
 
 	nonblock = blocking ? 0 : 1;
